Implement sound order functionality
- Add the methods for saving, retrieving and loading the sound order
- Disable reloading the current category when clicking the category in the SideBar

+++ UniversalSoundBoard/Common/ItemViewHolder.cs
@@ -665,7 +665,7 @@ namespace UniversalSoundBoard.Common
             get => _soundOrder;
             set
             {
-                localSettings.Values[soundOrderKey] = value;
+                localSettings.Values[soundOrderKey] = (int)value;
                 _soundOrder = value;
                 NotifyPropertyChanged("SoundOrder");
             }
